refactor: update dialogs_translator.py
Import Optional from typing for type checking purposes
Remove some comments that are not necessary(because the code is
cleaner now)

NEW: Modify try_translate_sentence return type

The first try-expect clause and return statement in
try_translate_sentence function is not nessarry so I removed it.
Instead of returning a tuple containing the translated text and success
flag, I change the function to return the translated text if it is
successful translated or return None if it failed to translate. I also
rewrite all the code cases in the translate function to adapt with the
new return type.